+++ contracts/zombiefactory.sol
@@ -9,15 +9,15 @@ contract ZombieFactory is Ownable {
     using SafeMath32 for uint32;
     using SafeMath16 for uint16;
 
-    event NewZombie(uint zombieId, string name, uint dna);
+    event NewZombie(uint256 zombieId, string name, uint256 dna);
 
-    uint dnaDigits = 16;
-    uint dnaModulus = 10 ** dnaDigits;
-    uint cooldownTime = 1 days;
+    uint256 dnaDigits = 16;
+    uint256 dnaModulus = 10 ** dnaDigits;
+    uint256 cooldownTime = 1 days;
 
     struct Zombie {
         string name;
-        uint dna;
+        uint256 dna;
         uint32 level;
         uint32 readyTime;
         uint16 winCount;
@@ -26,26 +26,26 @@ contract ZombieFactory is Ownable {
 
     Zombie[] public zombies;
 
-    mapping (uint => address) public zombieToOwner;
-    mapping (address => uint) ownerZombieCount;
+    mapping (uint256 => address) public zombieToOwner;
+    mapping (address => uint256) ownerZombieCount;
 
-    function _createZombie(string memory _name, uint _dna) internal {
-        uint id = zombies.push(Zombie(_name, _dna, 1, uint32(now + cooldownTime), 0, 0)) - 1;
+    function _createZombie(string memory _name, uint256 _dna) internal {
+        uint256 id = zombies.push(Zombie(_name, _dna, 1, uint32(now + cooldownTime), 0, 0)) - 1;
         /* msg is a global variable */
         zombieToOwner[id] = msg.sender;
         ownerZombieCount[msg.sender] = ownerZombieCount[msg.sender].add(1);
         emit NewZombie(id, _name, _dna);
     }
 
-    function _generateRandomDna(string memory _str) private view returns (uint) {
+    function _generateRandomDna(string memory _str) private view returns (uint256) {
         /* keccak is a SHA3 variant, it's use to hash a string */
-        uint rand = uint(keccak256(bytes(_str)));
+        uint256 rand = uint256(keccak256(bytes(_str)));
         return rand % dnaModulus;
     }
 
     function createRandomZombie(string memory _name) public {
         require(ownerZombieCount[msg.sender] == 0);
-        uint randDna = _generateRandomDna(_name);
+        uint256 randDna = _generateRandomDna(_name);
         _createZombie(_name, randDna);
     }
 }
